Move types into a singular types file
